Conventional Commit Message feat: Implement loading state for ClientsTable
This commit introduces a loading state to the ClientsTable component. When the component is first rendered, it will display a loading message while the data is being fetched. Once the data is available, the table will be displayed with the client information.

BREAKING CHANGE: The ClientsTable component now requires a data prop to be passed to it. This prop should be an array of client objects.

Additional notes:

The loading message can be customized to your liking.
You can also add a spinner or other visual indicator to show that the data is being loaded.
Make sure to import the getClients function from your ClientsService file.
Corte 12:00 PM.